% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempconv.R
\name{celcius_from_fahrenheit}
\alias{celcius_from_fahrenheit}
\title{Convert fahrenheit to celcius}
\usage{
celcius_from_fahrenheit(f)
}
\arguments{
\item{c}{Number, temperature in fahrenheit}
}
\value{
Temperature in celcius
}
\description{
Convert fahrenheit to celcius
}
