% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempconv.R
\name{fahrenheit_from_celcius}
\alias{fahrenheit_from_celcius}
\title{Convert celcius to fahrenheit}
\usage{
fahrenheit_from_celcius(c)
}
\arguments{
\item{c}{Number, temperature in celcius}
}
\value{
Temperature in Fahrenheit
}
\description{
Convert celcius to fahrenheit
}
